Separate into several logic classes.
