+++ projects/maxDankow/src/main/java/ru/mipt/diht/students/maxDankow/TwitterStream/ComandLineArgumentsParser.java
@@ -1,28 +1,30 @@
 package ru.mipt.diht.students.maxDankow.TwitterStream;
-import com.beust.jcommander.Parameter;
-import java.util.List;
-import java.util.ArrayList;
 
-public class JComanderArgsList {
+import com.beust.jcommander.Parameter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ComandLineArgumentsParser {
     private static final int DEFAULT_TWEETS_NUMBER_LIMIT = 100;
 
     @Parameter
     private List<String> parameters = new ArrayList<>();
 
-    @Parameter(names = { "--query", "-q" }, description = "Query or keywords to search")
+    @Parameter(names = {"--query", "-q"}, description = "Query or keywords to search")
     private String queryText;
 
-    @Parameter(names = { "--place", "-p" }, description = "Location of search (or nearby - near your position)")
-    private String locationStr = null;
+    @Parameter(names = {"--place", "-p"}, description = "Location of search")
+    private String locationName = null;
 
-    @Parameter(names = { "--stream", "-s" }, description = "Shows new tweets every second")
+    @Parameter(names = {"--stream", "-s"}, description = "Shows new tweets every second")
     private boolean streamMode = false;
 
-    @Parameter(names = { "--hideRetweets" }, description = "Hides retweets")
+    @Parameter(names = {"--hideRetweets"}, description = "Hides retweets")
     private boolean hideRetweets = false;
 
-    @Parameter(names = { "--limit", "-l" }, description = "Sets a limit to number of shown tweet")
-    private int tweetNumberLimit = DEFAULT_TWEETS_NUMBER_LIMIT;
+    @Parameter(names = {"--limit", "-l"}, description = "Sets a limit to number of shown tweets")
+    private int tweetsNumberLimit = DEFAULT_TWEETS_NUMBER_LIMIT;
 
     @Parameter(names = {"-h", "--help"}, description = "Print this page and exit", help = true)
     private boolean help = false;
@@ -31,8 +33,8 @@ public class JComanderArgsList {
         return queryText;
     }
 
-    public final String getLocationStr() {
-        return locationStr;
+    public final String getLocationName() {
+        return locationName;
     }
 
     public final boolean isStreamMode() {
@@ -44,11 +46,10 @@ public class JComanderArgsList {
     }
 
     public final int getTweetsNumberLimit() {
-        return tweetNumberLimit;
+        return tweetsNumberLimit;
     }
 
     public final boolean isHelp() {
         return help;
-
     }
 }
