+++ projects/maxDankow/src/test/java/ru/mipt/diht/students/maxDankow/TwitterStream/TwitterStreamTests.java
@@ -1,0 +1,5 @@
+package ru.mipt.diht.students.maxDankow.TwitterStream;
+
+public class TwitterStreamTests {
+
+}
